requirements.txt
pandas
numpy
notebook
scikit-learn
matplotlib

xgboost
catboost

python-box==6.0.2
PyYAML
types-PyYAML
ensure==1.0.2
joblib

Flask
Flask-Cors

-e .
